buildGOmap {clusterProfiler}	R Documentation

buildGOmap

Description

building GO mapping files

Usage

buildGOmap(gomap)

Arguments

gomap
data.frame with two columns of GO and gene ID

Details

provided by a data.frame of GO (column 1) and gene (column 2) direct annotation this function will building gene to GO and GO to gene mapping, with directly and undirectly (ancestor GO term) annotation.

Value

data.frame, GO annotation with indirect annotation

Author(s)

Yu Guangchuang

[Package clusterProfiler version 4.2.2 Index]