compareCluster {clusterProfiler}	R Documentation

Compare gene clusters functional profile

Description

Given a list of gene set, this function will compute profiles of each gene cluster.

Usage

compareCluster(geneClusters, fun = "enrichGO", data = "", ...)

Arguments

geneClusters
a list of entrez gene id. Alternatively, a formula of type Entrez~group

fun
One of "groupGO", "enrichGO", "enrichKEGG", "enrichDO" or "enrichPathway" .

data
if geneClusters is a formula, the data from which the clusters must be extracted.

...
Other arguments.

Value

A clusterProfResult instance.

Author(s)

Guangchuang Yu https://guangchuangyu.github.io

See Also

compareClusterResult-class, groupGO enrichGO

Examples

## Not run: 
data(gcSample)
xx <- compareCluster(gcSample, fun="enrichKEGG",
                     organism="hsa", pvalueCutoff=0.05)
as.data.frame(xx)
# plot(xx, type="dot", caption="KEGG Enrichment Comparison")

## formula interface
mydf <- data.frame(Entrez=c('1', '100', '1000', '100101467',
                            '100127206', '100128071'),
                   group = c('A', 'A', 'A', 'B', 'B', 'B'),
                   othergroup = c('good', 'good', 'bad', 'bad', 'good', 'bad'))
xx.formula <- compareCluster(Entrez~group, data=mydf,
                             fun='groupGO', OrgDb='org.Hs.eg.db')
as.data.frame(xx.formula)

## formula interface with more than one grouping variable
xx.formula.twogroups <- compareCluster(Entrez~group+othergroup, data=mydf,
                                       fun='groupGO', OrgDb='org.Hs.eg.db')
as.data.frame(xx.formula.twogroups)

## End(Not run)

[Package clusterProfiler version 4.2.2 Index]