enrichDAVID {clusterProfiler}	R Documentation

enrichDAVID

Description

enrichment analysis by DAVID

Usage

enrichDAVID(
  gene,
  idType = "ENTREZ_GENE_ID",
  universe,
  minGSSize = 10,
  maxGSSize = 500,
  annotation = "GOTERM_BP_FAT",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  qvalueCutoff = 0.2,
  species = NA,
  david.user
)

Arguments

gene
input gene

idType
id type

universe
background genes. If missing, the all genes listed in the database (eg TERM2GENE table) will be used as background.

minGSSize
minimal size of genes annotated for testing

maxGSSize
maximal size of genes annotated for testing

annotation
david annotation

pvalueCutoff
adjusted pvalue cutoff on enrichment tests to report

pAdjustMethod
one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"

qvalueCutoff
qvalue cutoff on enrichment tests to report as significant. Tests must pass i) pvalueCutoff on unadjusted pvalues, ii) pvalueCutoff on adjusted pvalues and iii) qvalueCutoff on qvalues to be reported.

species
species

david.user
david user

Value

A enrichResult instance

Author(s)

Guangchuang Yu

[Package clusterProfiler version 4.2.2 Index]