enrichGO {clusterProfiler}	R Documentation

GO Enrichment Analysis of a gene set. Given a vector of genes, this function will return the enrichment GO categories after FDR control.

Description

GO Enrichment Analysis of a gene set. Given a vector of genes, this function will return the enrichment GO categories after FDR control.

Usage

enrichGO(
  gene,
  OrgDb,
  keyType = "ENTREZID",
  ont = "MF",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe,
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500,
  readable = FALSE,
  pool = FALSE
)

Arguments

gene
a vector of entrez gene id.

OrgDb
OrgDb

keyType
keytype of input gene

ont
One of "BP", "MF", and "CC" subontologies, or "ALL" for all three.

pvalueCutoff
adjusted pvalue cutoff on enrichment tests to report

pAdjustMethod
one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"

universe
background genes. If missing, the all genes listed in the database (eg TERM2GENE table) will be used as background.

qvalueCutoff
qvalue cutoff on enrichment tests to report as significant. Tests must pass i) pvalueCutoff on unadjusted pvalues, ii) pvalueCutoff on adjusted pvalues and iii) qvalueCutoff on qvalues to be reported.

minGSSize
minimal size of genes annotated by Ontology term for testing.

maxGSSize
maximal size of genes annotated for testing

readable
whether mapping gene ID to gene Name

pool
If ont='ALL', whether pool 3 GO sub-ontologies

Value

An enrichResult instance.

Author(s)

Guangchuang Yu https://guangchuangyu.github.io

See Also

enrichResult-class, compareCluster

Examples

## Not run: 
  data(geneList, package = "DOSE")
	de <- names(geneList)[1:100]
	yy <- enrichGO(de, 'org.Hs.eg.db', ont="BP", pvalueCutoff=0.01)
	head(yy)

## End(Not run)

[Package clusterProfiler version 4.2.2 Index]