enrichKEGG {clusterProfiler}	R Documentation

KEGG Enrichment Analysis of a gene set. Given a vector of genes, this function will return the enrichment KEGG categories with FDR control.

Description

KEGG Enrichment Analysis of a gene set. Given a vector of genes, this function will return the enrichment KEGG categories with FDR control.

Usage

enrichKEGG(
  gene,
  organism = "hsa",
  keyType = "kegg",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe,
  minGSSize = 10,
  maxGSSize = 500,
  qvalueCutoff = 0.2,
  use_internal_data = FALSE
)

Arguments

gene
a vector of entrez gene id.

organism
supported organism listed in 'http://www.genome.jp/kegg/catalog/org_list.html'

keyType
one of "kegg", 'ncbi-geneid', 'ncib-proteinid' and 'uniprot'

pvalueCutoff
adjusted pvalue cutoff on enrichment tests to report

pAdjustMethod
one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"

universe
background genes. If missing, the all genes listed in the database (eg TERM2GENE table) will be used as background.

minGSSize
minimal size of genes annotated by Ontology term for testing.

maxGSSize
maximal size of genes annotated for testing

qvalueCutoff
qvalue cutoff on enrichment tests to report as significant. Tests must pass i) pvalueCutoff on unadjusted pvalues, ii) pvalueCutoff on adjusted pvalues and iii) qvalueCutoff on qvalues to be reported.

use_internal_data
logical, use KEGG.db or latest online KEGG data

Value

A enrichResult instance.

Author(s)

Guangchuang Yu http://guangchuangyu.github.io

See Also

enrichResult-class, compareCluster

Examples

## Not run: 
  data(geneList, package='DOSE')
  de <- names(geneList)[1:100]
  yy <- enrichKEGG(de, pvalueCutoff=0.01)
  head(yy)

## End(Not run)

[Package clusterProfiler version 4.2.2 Index]