enrichMKEGG {clusterProfiler}	R Documentation

KEGG Module Enrichment Analysis of a gene set. Given a vector of genes, this function will return the enrichment KEGG Module categories with FDR control.

Description

KEGG Module Enrichment Analysis of a gene set. Given a vector of genes, this function will return the enrichment KEGG Module categories with FDR control.

Usage

enrichMKEGG(
  gene,
  organism = "hsa",
  keyType = "kegg",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe,
  minGSSize = 10,
  maxGSSize = 500,
  qvalueCutoff = 0.2
)

Arguments

gene
a vector of entrez gene id.

organism
supported organism listed in 'http://www.genome.jp/kegg/catalog/org_list.html'

keyType
one of "kegg", 'ncbi-geneid', 'ncib-proteinid' and 'uniprot'

pvalueCutoff
adjusted pvalue cutoff on enrichment tests to report

pAdjustMethod
one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"

universe
background genes. If missing, the all genes listed in the database (eg TERM2GENE table) will be used as background.

minGSSize
minimal size of genes annotated by Ontology term for testing.

maxGSSize
maximal size of genes annotated for testing

qvalueCutoff
qvalue cutoff on enrichment tests to report as significant. Tests must pass i) pvalueCutoff on unadjusted pvalues, ii) pvalueCutoff on adjusted pvalues and iii) qvalueCutoff on qvalues to be reported.

Value

A enrichResult instance.

[Package clusterProfiler version 4.2.2 Index]