enricher {clusterProfiler}	R Documentation

enricher

Description

A universal enrichment analyzer

Usage

enricher(
  gene,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe,
  minGSSize = 10,
  maxGSSize = 500,
  qvalueCutoff = 0.2,
  TERM2GENE,
  TERM2NAME = NA
)

Arguments

gene
a vector of gene id

pvalueCutoff
adjusted pvalue cutoff on enrichment tests to report

pAdjustMethod
one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"

universe
background genes. If missing, the all genes listed in the database (eg TERM2GENE table) will be used as background.

minGSSize
minimal size of genes annotated for testing

maxGSSize
maximal size of genes annotated for testing

qvalueCutoff
qvalue cutoff on enrichment tests to report as significant. Tests must pass i) pvalueCutoff on unadjusted pvalues, ii) pvalueCutoff on adjusted pvalues and iii) qvalueCutoff on qvalues to be reported.

TERM2GENE
user input annotation of TERM TO GENE mapping, a data.frame of 2 column with term and gene

TERM2NAME
user input of TERM TO NAME mapping, a data.frame of 2 column with term and name

Value

A enrichResult instance

Author(s)

Guangchuang Yu

[Package clusterProfiler version 4.2.2 Index]