groupGO {clusterProfiler}	R Documentation

Functional Profile of a gene set at specific GO level. Given a vector of genes, this function will return the GO profile at a specific level.

Description

Functional Profile of a gene set at specific GO level. Given a vector of genes, this function will return the GO profile at a specific level.

Usage

groupGO(
  gene,
  OrgDb,
  keyType = "ENTREZID",
  ont = "CC",
  level = 2,
  readable = FALSE
)

Arguments

gene
a vector of entrez gene id.

OrgDb
OrgDb

keyType
key type of input gene

ont
One of "MF", "BP", and "CC" subontologies.

level
Specific GO Level.

readable
if readable is TRUE, the gene IDs will mapping to gene symbols.

Value

A groupGOResult instance.

Author(s)

Guangchuang Yu http://guangchuangyu.github.io

See Also

groupGOResult-class, compareCluster

Examples

	data(gcSample)
	yy <- groupGO(gcSample[[1]], 'org.Hs.eg.db', ont="BP", level=2)
	head(summary(yy))
	#plot(yy)

[Package clusterProfiler version 4.2.2 Index]