groupGOResult-class {clusterProfiler}	R Documentation

Class "groupGOResult" This class represents the result of functional Profiles of a set of gene at specific GO level.

Description

Class "groupGOResult" This class represents the result of functional Profiles of a set of gene at specific GO level.

Slots

result

GO classification result

ontology

Ontology

level

GO level

organism

one of "human", "mouse" and "yeast"

gene

Gene IDs

readable

logical flag of gene ID in symbol or not.

Author(s)

Guangchuang Yu https://guangchuangyu.github.io

See Also

compareClusterResult compareCluster groupGO

[Package clusterProfiler version 4.2.2 Index]