gseWP {clusterProfiler}	R Documentation

gseWP

Description

GSEA analysis for WikiPathways

Usage

gseWP(geneList, organism, ...)

Arguments

geneList
ranked gene list

organism
supported organisms, which can be accessed via the get_wp_organisms() function

...
additional parameters, see also the parameters supported by the enricher() function

Details

This function performs GSEA using WikiPathways

Value

A gseaResult instance

Author(s)

Guangchuang Yu

[Package clusterProfiler version 4.2.2 Index]